Add "Action in progress" error for multi-engine.
Raises an "Action in progress" error if an action is in progress when
another action is issued.  The engine confirms the "in progress" status
by sending an RPC to the engine doing the work.  If the stack's status
is "in progress", but no response is received, the other engine is
assumed to have died and the action will proceed as usual.

Blueprint multi-engine
Fixes bug #1211276

Change-Id: I9f03a2281e5dea04e45fec61fbdb1120f0a32ef0