Make Server compatible with Nova Key Pair resource
The Server resource checks that the specified key name exists in
validate(), but this will always fail if the specified key is created
with via a Nova Key Pair resource.

Add a check for a Nova Key Pair resource with a name property
corresponding to the Server's key_name property during SSH key
validation in Server.validate().  This will allow the Server resource to
use a Nova Key Pair-generated SSH key.

Change-Id: I1cd6aa5c6555657c376df702754fd3d270ac44ec
Closes-Bug: #1271303