Add Resource to use Rackspace RackConnect
Add a resource which can be used in Rackspace's Public cloud to
add servers to a RackConnected load balancer pool and assign
public IP's to the added server. Boot From Volume servers seem
to have some delay in becoming visible to rackconnect. Therefore,
a check is introduced in the resource to try repeatedly until the
rackconnected resource is created.

Handle not found on initial deletes

Change-Id: I05624450444a1cc211fc2d2c77a20be354be04b8