Make heat-api work with keystone v2
This is required to run a local instance of Heat that talks to RS cloud.
It's not necessary in our dev/prod environments because Repose does the
authentication.

I tried making this work with keystone v2/v3 so that we can submit this
fix upstream by using:
http://docs.openstack.org/developer/python-keystoneclient/api/keystoneclient.auth.identity.generic.html

...but the "version independent plugins" framework is not fully
implemented:
https://github.com/openstack/python-keystoneclient/blob/841e634acea/keystoneclient/auth/identity/generic/base.py#L75-L92

I will talk to Jamie Lennox to see if there's any way to implement this
as a generic solution.

Change-Id: I35cd2ff77fb0bcf5fd90628b4fd58514f61fbae0